+++ base/base.jl
@@ -162,9 +162,9 @@ end
 
 finalize(o::ANY) = ccall(:jl_finalize, Void, (Any,), o)
 
-gc(full = true) = ccall(:jl_gc_collect, Void, (Int,), full ? 1 : 0)
-gc_enable() = ccall(:jl_gc_enable, Void, ())
-gc_disable() = ccall(:jl_gc_disable, Void, ())
+gc(full::Bool=true) = ccall(:jl_gc_collect, Void, (Cint,), full)
+gc_enable() = bool(ccall(:jl_gc_enable, Cint, ()))
+gc_disable() = bool(ccall(:jl_gc_disable, Cint, ()))
 
 bytestring(str::ByteString) = str